+++ tests/test_by/forms.py
@@ -1,0 +1,10 @@
+from django.forms import ModelForm
+
+from .models import BYTestModel
+
+
+class BYTestForm(ModelForm):
+
+    class Meta:
+        model = BYTestModel
+        fields = ('region', 'pass_num', 'pass_id', 'postal_code')