+++ docs/changelog.rst
@@ -8,6 +8,9 @@ New flavors:
 
 - Nepal LocalFlavor: Support for Nepal added
   (`gh-451 <https://github.com/django/django-localflavor/pull/451>`_).
+- Belarus localflavor
+  (`gh-422 <https://github.com/django/django-localflavor/pull/422>`_,
+  `gh-442 <https://github.com/django/django-localflavor/pull/442>`_).
 
 New fields for existing flavors: